refortor: 게임 종료 기능 분리 및 유닛 테스트 클래시 단위로 분리
